+++ stories/components/SInputDropdown.02_Avatar_Options.story.vue
@@ -17,7 +17,7 @@ const options: OptionAvatar[] = [
 
 function initState() {
   return {
-    size: 'small'
+    size: 'md'
   }
 }
 </script>
@@ -28,9 +28,8 @@ function initState() {
       <HstSelect
         title="size"
         :options="{
-          mini: 'mini',
-          small: 'small',
-          medium: 'medium'
+          sm: 'sm',
+          md: 'md'
         }"
         v-model="state.size"
       />