+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [4.29.1](https://github.com/globalbrain/sefirot/compare/v4.29.0...v4.29.1) (2025-07-25)
+
+
+### Bug Fixes
+
+* **input:** bring back `mini` size ([f1b2cd6](https://github.com/globalbrain/sefirot/commit/f1b2cd6adfaaf2a4b448806746d2af6f6b1a5ba7))
+
+# [4.29.0](https://github.com/globalbrain/sefirot/compare/v4.28.0...v4.29.0) (2025-07-25)
+
+
+### Features
+
+* **button,input:** add new `sm` and `md` size ([#637](https://github.com/globalbrain/sefirot/issues/637)) ([14f855c](https://github.com/globalbrain/sefirot/commit/14f855c38ccd07946ea497be4e9a06d747935260))
+* **content:** refine content styles ([#636](https://github.com/globalbrain/sefirot/issues/636)) ([4e1a8b4](https://github.com/globalbrain/sefirot/commit/4e1a8b45c02a979213f836917ad0fdb53c5de2e9))
+
 # [4.28.0](https://github.com/globalbrain/sefirot/compare/v4.27.0...v4.28.0) (2025-07-22)